Modificate structure
