+++ moneta/db/moneta/data/image.sql
@@ -1,22 +1,22 @@
-INSERT INTO image (id, link)
+INSERT INTO image (id, css, category)
 VALUES
-    (1, link),
-    (2, link),
-    (3, link),
-    (4, link),
-    (5, link),
-    (6, link),
-    (7, link),
-    (8, link),
-    (9, link),
-    (10, link),
-    (11, link),
-    (12, link),
-    (13, link),
-    (14, link),
-    (15, link),
-    (16, link),
-    (17, link),
-    (18, link),
-    (19, link),
-    (20, link);
+    (1, 'coins', 'current'),
+    (2, 'coin-9', 'current'),
+    (3, 'credit-card-6', 'current'),
+    (4, 'notes', 'current'),
+    (5, 'notes-2', 'current'),
+    (6, 'piggy-bank-1', 'current'),
+    (7, 'safebox-4', 'current'),
+    (8, 'wallet-1', 'current'),
+    (9, 'asd', 'expend'),
+    (10, 'asd', 'expend'),
+    (11, 'asd', 'expend'),
+    (12, 'asd', 'expend'),
+    (13, 'asd', 'expend'),
+    (14, 'asd', 'income'),
+    (15, 'asd', 'income'),
+    (16, 'asd', 'income'),
+    (17, 'asd', 'income'),
+    (18, 'asd', 'income'),
+    (19, 'asd', 'income'),
+    (20, 'asd', 'income');
