+++ moneta/db/moneta/data/income.sql
@@ -1,22 +1,22 @@
-INSERT INTO income (id, name, currency, user_id, create_time, mod_time, image_id)
+INSERT INTO income (id, name, currency, user_id, create_time, mod_time, amount, image_id)
 VALUES
-    (1, 'ataxies', 'USD', 20, 1551244048.256399, NULL, 1),
-    (2, 'tyste', 'UAH', 4, 1549271862.256399, NULL, 13),
-    (3, 'accrescence', 'EUR', 9, 1548359532.256399, NULL, 15),
-    (4, 'malto', 'UAH', 6, 1550183330.256399, NULL, 1),
-    (5, 'bullhorn', 'UAH', 13, 1547659366.256399, NULL, 5),
-    (6, 'visceromotor', 'USD', 11, 1546969782.256399, NULL, 13),
-    (7, 'aloelike', 'GBP', 9, 1549724059.256399, NULL, 18),
-    (8, 'carbamidine', 'EUR', 5, 1545695898.256399, NULL, 2),
-    (9, 'chopa', 'UAH', 14, 1552360250.256399, NULL, 9),
-    (10, 'rudistae', 'EUR', 11, 1545719185.256399, NULL, 1),
-    (11, 'dowp', 'UAH', 7, 1547791162.256399, NULL, 13),
-    (12, 'quadrigenarious', 'EUR', 15, 1549521425.256399, NULL, 9),
-    (13, 'vacantia', 'UAH', 10, 1547311651.256399, NULL, 6),
-    (14, 'mancipia', 'EUR', 15, 1548299329.256399, NULL, 1),
-    (15, 'nonpossible', 'USD', 12, 1552177752.256399, NULL, 20),
-    (16, 'soldierliness', 'USD', 2, 1545914781.256399, NULL, 18),
-    (17, 'caracore', 'EUR', 19, 1543758451.256399, NULL, 18),
-    (18, 'olibanums', 'UAH', 1, 1551187165.256399, NULL, 3),
-    (19, 'goblinry', 'USD', 6, 1549051916.256399, NULL, 1),
-    (20, 'disproportionality', 'EUR', 3, 1552208092.256399, NULL, 17);
+    (1, 'ataxies', 'USD', 20, 1551244048.256399, NULL, 91731, 1),
+    (2, 'tyste', 'UAH', 4, 1549271862.256399, NULL, 91110, 13),
+    (3, 'accrescence', 'EUR', 9, 1548359532.256399, NULL, 10280, 15),
+    (4, 'malto', 'UAH', 6, 1550183330.256399, NULL, 43142, 1),
+    (5, 'bullhorn', 'UAH', 13, 1547659366.256399, NULL, 72598, 5),
+    (6, 'visceromotor', 'USD', 11, 1546969782.256399, NULL, 29437, 13),
+    (7, 'aloelike', 'GBP', 9, 1549724059.256399, NULL, 47343, 18),
+    (8, 'carbamidine', 'EUR', 5, 1545695898.256399, NULL, 55295, 2),
+    (9, 'chopa', 'UAH', 14, 1552360250.256399, NULL, 35969, 9),
+    (10, 'rudistae', 'EUR', 11, 1545719185.256399, NULL, 91863, 1),
+    (11, 'dowp', 'UAH', 7, 1547791162.256399, NULL, 94013, 13),
+    (12, 'quadrigenarious', 'EUR', 15, 1549521425.256399, NULL, 99404, 9),
+    (13, 'vacantia', 'UAH', 10, 1547311651.256399, NULL, 51971, 6),
+    (14, 'mancipia', 'EUR', 15, 1548299329.256399, NULL, 9498, 1),
+    (15, 'nonpossible', 'USD', 12, 1552177752.256399, NULL, 97576, 20),
+    (16, 'soldierliness', 'USD', 2, 1545914781.256399, NULL, 21082, 18),
+    (17, 'caracore', 'EUR', 19, 1543758451.256399, NULL, 99485, 18),
+    (18, 'olibanums', 'UAH', 1, 1551187165.256399, NULL, 32545, 3),
+    (19, 'goblinry', 'USD', 6, 1549051916.256399, NULL, 83233, 1),
+    (20, 'disproportionality', 'EUR', 3, 1552208092.256399, NULL, 17100, 17);
