+++ moneta/db/moneta/data/user.sql
@@ -1,22 +1,22 @@
-INSERT INTO user (id, password, def_currency, email, is_activated)
+INSERT INTO user (id, def_currency, is_activated)
 VALUES
-    (1, 'theomorphism', 'UAH', 'undanceable@gmail.com', 1),
-    (2, 'unimprovable', 'GBP', 'lannerets@gmail.com', 1),
-    (3, 'skilts', 'GBP', 'rheinic@gmail.com', 1),
-    (4, 'attriteness', 'USD', 'unbribing@gmail.com', 1),
-    (5, 'pyrenodeous', 'EUR', 'homopiperonyl@gmail.com', 1),
-    (6, 'poleaxer', 'UAH', 'caramboling@gmail.com', 1),
-    (7, 'dermonecrotic', 'GBP', 'unslockened@gmail.com', 1),
-    (8, 'extrovertively', 'USD', 'trippant@gmail.com', 1),
-    (9, 'laramie', 'USD', 'cocain@gmail.com', 1),
-    (10, 'genies', 'UAH', 'nailfile@gmail.com', 1),
-    (11, 'neighborhoods', 'USD', 'shrinkages@gmail.com', 1),
-    (12, 'paravant', 'UAH', 'southwood@gmail.com', 1),
-    (13, 'gelant', 'EUR', 'pithecan@gmail.com', 1),
-    (14, 'deeses', 'GBP', 'styful@gmail.com', 1),
-    (15, 'postphrenic', 'GBP', 'diglottism@gmail.com', 1),
-    (16, 'jant', 'GBP', 'falcation@gmail.com', 1),
-    (17, 'shagbarks', 'GBP', 'mesteno@gmail.com', 1),
-    (18, 'macacos', 'UAH', 'overgratifying@gmail.com', 0),
-    (19, 'meguilp', 'GBP', 'anterethic@gmail.com', 0),
-    (20, 'pseudocele', 'USD', 'rebalance@gmail.com', 0);
+    (1, 'UAH', 1),
+    (2, 'GBP', 1),
+    (3, 'GBP', 1),
+    (4, 'USD', 1),
+    (5, 'EUR', 1),
+    (6, 'UAH', 1),
+    (7, 'GBP', 1),
+    (8, 'USD', 1),
+    (9, 'USD', 1),
+    (10, 'UAH', 1),
+    (11, 'USD', 1),
+    (12, 'UAH', 1),
+    (13, 'EUR', 1),
+    (14, 'GBP', 1),
+    (15, 'GBP', 1),
+    (16, 'GBP', 1),
+    (17, 'GBP', 1),
+    (18, 'UAH', 0),
+    (19, 'GBP', 0),
+    (20, 'USD', 0);
