+++ moneta/db/moneta/schema/current.sql
@@ -5,7 +5,6 @@ CREATE TABLE IF NOT EXISTS db_moneta.current (
   id INT NOT NULL AUTO_INCREMENT,
   name VARCHAR(45) NULL,
   currency VARCHAR(45) NULL,
-  is_include TINYINT(1) NULL,
   create_time INT(11) NULL,
   mod_time INT(11) NULL,
   amount FLOAT NULL,
@@ -14,5 +13,6 @@ CREATE TABLE IF NOT EXISTS db_moneta.current (
   CONSTRAINT fk_current_image
     FOREIGN KEY (image_id)
     REFERENCES image (id)
+    ON DELETE CASCADE
 )
 ENGINE = InnoDB CHARSET=utf8;
