+++ moneta/db/moneta/schema/expend.sql
@@ -6,12 +6,13 @@ CREATE TABLE IF NOT EXISTS db_moneta.expend (
   name VARCHAR(45) NULL,
   currency VARCHAR(45) NOT NULL,
   create_time INT(11) NULL,
-  modification_time INT(11) NULL,
-  planned_cost FLOAT NULL,
+  mod_time INT(11) NULL,
+  amount FLOAT NULL,
   image_id INT NOT NULL,
   PRIMARY KEY (id),
   CONSTRAINT fk_expend_image
     FOREIGN KEY (image_id)
     REFERENCES image (id)
+    ON DELETE CASCADE
 )
 ENGINE = InnoDB  CHARSET=utf8;
