+++ moneta/db/moneta/schema/image.sql
@@ -1,9 +1,10 @@
 /*
-    Table holds links to images.
+    Table holds links to images and .
 */
 CREATE TABLE IF NOT EXISTS db_moneta.image (
-  id INT NOT NULL,
-  link VARCHAR(45) NOT NULL,
+  id INT NOT NULL AUTO_INCREMENT,
+  css VARCHAR(45) NOT NULL,
+  category VARCHAR(45) NOT NULL,
   PRIMARY KEY (id)
 )
 ENGINE = InnoDB CHARSET=utf8;
