--- moneta/src/python/current/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class CurrentAppConfig(AppConfig):
-    name = 'current'
