--- moneta/src/python/customauth/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class CustomAuthAppConfig(AppConfig):
-    name = 'customauth'
