--- moneta/src/python/customauth/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from . import views
-
-
-urlpatterns = [
-    path('', views.home, name='moneta-home'),
-
-]
