--- moneta/src/python/expend/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ExpendAppConfig(AppConfig):
-    name = 'expend'
