--- moneta/src/python/income/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class IncomeAppConfig(AppConfig):
-    name = 'income'
