+++ moneta/www/settings/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for moneta project.
+Django settings for settings project.
 
 Generated by 'django-admin startproject' using Django 2.1.7.
 
@@ -9,23 +9,31 @@ https://docs.djangoproject.com/en/2.1/topics/settings/
 For the full list of settings and their values, see
 https://docs.djangoproject.com/en/2.1/ref/settings/
 """
+
 import os
+
+
+# Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-SECRET_KEY = 'xxv_7j77l%31%a*n5ac606(%gok5krdtc0x(9e275#^_yea6dv'
-DEBUG = True
-ALLOWED_HOSTS = []
+
 
 STATIC_URL = '/static/'
 
-STATICFILES_DIR = [
+STATICFILES_DIRS = [
     os.path.join(BASE_DIR, "static")
 ]
-STATIC_ROOT = os.path.join(BASE_DIR, 'staticfiles')
-STATICFILES_FINDERS = (
-    'django.contrib.staticfiles.finders.FileSystemFinder',
-    'django.contrib.staticfiles.finders.AppDirectoriesFinder',
 
-)
+# Quick-start development settings - unsuitable for production
+# See https://docs.djangoproject.com/en/2.1/howto/deployment/checklist/
+
+# SECURITY WARNING: keep the secret key used in production secret!
+SECRET_KEY = 'xxv_7j77l%31%a*n5ac606(%gok5krdtc0x(9e275#^_yea6dv'
+
+# SECURITY WARNING: don't run with debug turned on in production!
+DEBUG = True
+
+ALLOWED_HOSTS = []
+
 
 # Application definition
 
@@ -35,14 +43,6 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'crispy_forms',
-    'customauth',
-    'current',
-    'income',
-    'expend',
-    'helper',
-    'tests',
-    'authentication'
 ]
 
 MIDDLEWARE = [
@@ -55,13 +55,13 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'moneta.urls'
+ROOT_URLCONF = 'www.settings.urls'
 
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [os.path.join(os.path.dirname(os.path.dirname(BASE_DIR)), 'www', 'templates')],
-        'APP_DIRS': True,
+        'DIRS': [os.path.join(BASE_DIR, "templates")],
+        'APP_DIRS': False,
         'OPTIONS': {
             'context_processors': [
                 'django.template.context_processors.debug',
@@ -73,15 +73,7 @@ TEMPLATES = [
     },
 ]
 
-STATIC_URL = '/static/'
-
-STATICFILES_DIRS = [
-    os.path.join(os.path.dirname(os.path.dirname(BASE_DIR)), 'www', 'static'),
-]
-
-LOGIN_URL = '/login'
-
-WSGI_APPLICATION = 'moneta.wsgi.application'
+WSGI_APPLICATION = 'www.settings.wsgi.application'
 
 #Configurations to send email
 
@@ -100,12 +92,10 @@ DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.mysql',
         'OPTIONS': {
-            'read_default_file': '/etc/mysql/moneta.cnf',
-
+            'read_default_file': os.path.join(os.path.dirname(BASE_DIR), 'etc') + '/moneta_mysql.cnf',
         },
     }
 }
-
 
 # Password validation
 # https://docs.djangoproject.com/en/2.1/ref/settings/#auth-password-validators
@@ -144,11 +134,6 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/2.1/howto/static-files/
 
 STATIC_URL = '/static/'
-
-# try:
-#     from .local_settings import *
-# except ImportError:
-#     pass
 
 
 try:
