--- moneta/www/templates/home.html
@@ -1,31 +1,0 @@
-{% extends "base.html" %}
-{% load crispy_forms_tags %}
-{% block content %}
-<div>
-
-    <div class="container">
-        <!--Div for Income--->
-        <h5>Income</h5>
-        <div class="child-container">
-            <button class="button">First</button>
-            <button class="buttonn">+</button>
-        </div>
-    </div>
-    <div class="container">
-        <!--Div for Current--->
-        <h5>Current</h5>
-        <div class="child-container">
-            <button class="button">First</button>
-            <button class="buttonn">+</button>
-        </div>
-    </div>
-    <div class="container">
-        <!--Div for Expend--->
-        <h5>Expend</h5>
-        <div class="child-container">
-            <button class="button">First</button>
-            <button class="buttonn">+</button>
-        </div>
-    </div>
-</div>
-{% endblock content %}
