--- moneta/www/templates/login_app/logout.html
@@ -1,9 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<h2>You have been logged out</h2>
-<div class="border-top pt-3">
-    <small class="text-muted">
-        <a href="{% url 'login' %}">Log In Again</a>
-    </small>
-</div>
-{% endblock content %}
